move v2 initialize logic to new lib; switch to inat fork of hapi-joi-to-swagger; add map tile grid and place endpoints; fix users update
