+++ lib/controllers/v1/observations_controller.js
@@ -1120,7 +1120,7 @@ ObservationsController.observationsSpeciesObserverCounts = async req => {
 };
 
 // returns unviewed notifications about new IDs and comments
-// on the authroized user's observations
+// on the authorized user's observations
 ObservationsController.updates = async req => {
   if ( !req.userSession ) {
     throw new Error( 401 );
