+++ package.json
@@ -39,8 +39,8 @@
     "geoip-lite": "^1.3.4",
     "graphql": "*",
     "graphql-fields": "*",
-    "handlebars": "^4.5.3",
-    "hapi-joi-to-swagger": "github:pleary/hapi-joi-to-swagger#6c17eccd5a6f6e3b0a50646ad83a6459c11f8809",
+    "handlebars": "^4.7.6",
+    "hapi-joi-to-swagger": "github:inaturalist/hapi-joi-to-swagger#0ec177ccc9f8306f6fbcc7066174313e593c0600",
     "inaturalistjs": "github:inaturalist/inaturalistjs",
     "jsonwebtoken": "^8.3.0",
     "lodash": "^4.17.15",
